Update 산지니의 여행계획_solution.cpp

+++ 2023_CodeRace_제출본/고학년_산지니의 여행계획/산지니의 여행계획_solution.cpp
@@ -3,11 +3,11 @@
 
 using namespace std;
 
-int getParent(int num);
-void unionParent(int a, int b);
-bool isSameParent(int a, int b);
+long long getParent(long long num);
+void unionParent(long long a, long long b);
+bool isSameParent(long long a, long long b);
 
-void dfs(int node);
+void dfs(long long node);
 
 void fastio();
 void input();
@@ -16,9 +16,9 @@ void output();
 
 class Edge {
 public:
-    int node[2];
-    int dis;
-    Edge(int a, int b, int _dis) {
+    long long node[2];
+    long long dis;
+    Edge(long long a, long long b, long long _dis) {
         node[0] = a;
         node[1] = b;
         dis = _dis;
@@ -30,43 +30,43 @@ public:
 
 
 
-int N, M, S, answer = 0;
+long long N, M, S, answer = 0;
 
 vector<Edge> edge;
-vector<pair<int, int>> kruskalEdge[MAX];
-int parent[MAX];
+vector<pair<long long, long long>> kruskalEdge[MAX];
+long long parent[MAX];
 
 bool visited[MAX];
-int v[MAX];
+long long v[MAX];
 
 
 
-int getParent(int num){
+long long getParent(long long num){
     if (parent[num] == num) return num;
     return parent[num] = getParent(parent[num]);
 }
 
-void unionParent(int a, int b){
+void unionParent(long long a, long long b){
     a = getParent(a);
     b = getParent(b);
     if (a < b) parent[b] = a;
     else parent[a] = b;
 }
 
-bool isSameParent(int a, int b){
+bool isSameParent(long long a, long long b){
     return getParent(a) == getParent(b);
 }
 
 void Kruskal(){
     sort(edge.begin(), edge.end());
-    for (int i = 1; i <= N; i++) {
+    for (long long i = 1; i <= N; i++) {
         parent[i] = i;
     }
-    int sum = 0;
-    for (int i = 0; i < edge.size() ; i++){
-        int nodeA = edge[i].node[0];
-        int nodeB = edge[i].node[1];
-        int dis = edge[i].dis;
+    long long sum = 0;
+    for (long long i = 0; i < edge.size() ; i++){
+        long long nodeA = edge[i].node[0];
+        long long nodeB = edge[i].node[1];
+        long long dis = edge[i].dis;
         if (!isSameParent(nodeA, nodeB)){
             kruskalEdge[nodeA].push_back({nodeB, dis});
             kruskalEdge[nodeB].push_back({nodeA, dis});
@@ -76,11 +76,11 @@ void Kruskal(){
     }
 }
 
-void dfs(int node){
+void dfs(long long node){
     visited[node] = 1;
     for(auto next : kruskalEdge[node]){
-        int nextNode = next.first;
-        int nextDis = next.second;
+        long long nextNode = next.first;
+        long long nextDis = next.second;
         if(visited[nextNode]) continue;
         v[nextNode] += nextDis + v[node];
         dfs(nextNode);
@@ -94,8 +94,8 @@ void fastio(){
 
 void input(){
     cin >> N >> M;
-    for (int i = 0; i < M; i++) {
-        int s, e, dis;
+    for (long long i = 0; i < M; i++) {
+        long long s, e, dis;
         cin >> s >> e >> dis;
         edge.push_back(Edge(s, e, dis));
     }
@@ -106,7 +106,7 @@ void solution() {
     Kruskal();
     answer *= 2;
     dfs(S);
-    int lastTraverse = *max_element(v+1, v+N+1);
+    long long lastTraverse = *max_element(v+1, v+N+1);
     answer -= lastTraverse;
 }
